+++ app/src/main/java/alexander/voronov/marvelcharacters/domain/entities/Data.kt
@@ -1,4 +1,4 @@
-package alexander.voronov.marvelcharacters.entities
+package alexander.voronov.marvelcharacters.domain.entities
 
 data class Data(
     val count: String,
