+++ app/src/main/java/alexander/voronov/marvelcharacters/domain/entities/MyResponse.kt
@@ -1,0 +1,5 @@
+package alexander.voronov.marvelcharacters.domain.entities
+
+data class MyResponse(
+    val data: Data
+)
