+++ app/src/main/java/alexander/voronov/marvelcharacters/domain/entities/Result.kt
@@ -1,4 +1,4 @@
-package alexander.voronov.marvelcharacters.entities
+package alexander.voronov.marvelcharacters.domain.entities
 
 data class Result(
     val id: Long,
