--- app/src/main/java/alexander/voronov/marvelcharacters/entities/MyResponse.kt
@@ -1,5 +1,0 @@
-package alexander.voronov.marvelcharacters.entities
-
-data class MyResponse(
-    val data: Data
-)
