+++ app/src/main/java/alexander/voronov/marvelcharacters/ui/ItemViewHolder.kt
@@ -1,7 +1,8 @@
-package alexander.voronov.marvelcharacters
+package alexander.voronov.marvelcharacters.ui
 
+import alexander.voronov.marvelcharacters.R
 import alexander.voronov.marvelcharacters.databinding.ElementRecyclerviewBinding
-import alexander.voronov.marvelcharacters.entities.Result
+import alexander.voronov.marvelcharacters.domain.entities.Result
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
